server: Allows anonymous users access to ./well-known/jwks.json
The ./well-known/jwks.json endpoint contains important, publicly accessible
keys for validating signatures such as the OpenID Connect ID Token signature.

Currently, this endpoint shows the public key for validating ID Tokens
only. As this key is public, a policy was added which allows any user
(including anonymous ones) to access this specific key.

Thus, administrators no longer need to add a policy to allow access to
this endpoint on a fresh installation. It is still possible to change
this behaviour by removing the policy
("hydra policies delete default-oidc-id-token-public-policy") or
replacing it.

This change affects new installations only.

Closes #761