feat: support jwt signing alg other than RS256 (#1889)
Closes #1817

Add new field TokenEndpointAuthSigningAlgorithm for Client struct which by default will be returned as RS256 from the "getter" function GetTokenEndpointAuthSigningAlgorithm unless a defined value is set in client payload when performing a request to the Create OAuth 2.0 client endpoint. This of course implies a migration to be applied to the hydra_cient table to include this new field.