feat: Add all subjects option to list consent sessions endpoint
This adds `all` subjects option to list consent sessions endpoint: https://www.ory.sh/hydra/docs/reference/api#lists-all-consent-sessions-of-a-subject

This is to allow getting for all subjects explicitly instead of for a single subject.

It follows similar to `all` option for the endpoint to revoke consent sessions for a subject (for a specific client or all clients).

Regenerated swagger based code:
```
make sdk
```

Test passed:
```
make test
```