fix(oauth2): enforce assertion check on userinfo aud field
This is so the check on the `ok` variable is effectual. Prior to this patch the type assertion on the *client.Client was setting the value of `ok`. Due to the fact the type assertion on *client.Client is already checked and on a false value it exits the func, this value will *always* be true.

Signed-off-by: James Elliott <james-d-elliott@users.noreply.github.com>