feat: introduce the concept of a flow in the persistence layer
This commit merges the tables authentication_request and
authentication_request_handled into a new table named flow in order to simplify
and optimize the database schema. The changes are limited to the persistence
layer without affecting the API.

Only sqlite migrations are implemented. Migrations for other databases are to
be added later.